+++ packages/dumpling/src/platte.ts
@@ -1,4 +1,5 @@
 import { fromPlatte as fromButtonPlatte } from './button/platte'
+import { fromPlatte as fromSwitchPlatte } from './switch/platte'
 
 type FromPlatteFn = (color: string) => {
   name: string
@@ -7,6 +8,7 @@ type FromPlatteFn = (color: string) => {
 
 const fromPlatte: Record<string, FromPlatteFn> = {
   Button: fromButtonPlatte,
+  Switch: fromSwitchPlatte,
 }
 
 export default fromPlatte